backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
google-cloud-storage==2.14.0
google-api-python-client==2.170.0
google-auth-httplib2==0.2.0
google-auth==2.35.0
google-auth-oauthlib==1.2.0
python-multipart==0.0.17
pydantic==2.10.3
python-dotenv==1.0.1
motor==3.6.0
beanie==1.27.0
pymongo==4.9.0
openai==1.86.0
PyJWT==2.10.1
email-validator==2.2.0
pytest==8.3.4
pytest-asyncio==0.23.8
httpx==0.28.1
bcrypt==4.1.2 